+++ src/main/java/chess/controller/dto/board/MovablePathResponseDto.java
@@ -1,4 +1,4 @@
-package chess.controller.dto;
+package chess.controller.dto.board;
 
 import chess.domain.board.position.Path;
