+++ src/main/java/chess/controller/dto/board/PieceResponseDto.java
@@ -1,4 +1,4 @@
-package chess.controller.dto;
+package chess.controller.dto.board;
 
 public class PieceResponseDto {
