+++ src/main/java/chess/controller/dto/score/GameResultDto.java
@@ -1,4 +1,4 @@
-package chess.controller.dto;
+package chess.controller.dto.score;
 
 import chess.domain.manager.GameStatus;
 import chess.domain.piece.Owner;
