+++ src/main/java/chess/controller/dto/score/StatusResponseDto.java
@@ -1,4 +1,4 @@
-package chess.controller.dto;
+package chess.controller.dto.score;
 
 import chess.domain.manager.GameStatus;
