+++ src/main/java/chess/dao/ChessDao.java
@@ -1,6 +1,11 @@
 package chess.dao;
 
-import chess.controller.dto.*;
+import chess.controller.dto.board.PieceResponseDto;
+import chess.controller.dto.game.GameResponseDto;
+import chess.controller.dto.game.NewGameRequestDto;
+import chess.controller.dto.history.HistoryResponseDto;
+import chess.controller.dto.score.ScoreResponseDto;
+import chess.controller.dto.state.StateResponseDto;
 import chess.domain.manager.ChessManager;
 import chess.domain.manager.GameStatus;
 import chess.domain.piece.Piece;
