+++ src/main/java/chess/domain/command/Restart.java
@@ -1,4 +1,4 @@
-package chess.controller.command;
+package chess.domain.command;
 
 import chess.domain.manager.ChessManager;
