+++ src/main/java/chess/domain/command/Show.java
@@ -1,4 +1,4 @@
-package chess.controller.command;
+package chess.domain.command;
 
 import chess.domain.board.position.Path;
 import chess.domain.board.position.Position;
