+++ src/main/java/chess/domain/command/Status.java
@@ -1,4 +1,4 @@
-package chess.controller.command;
+package chess.domain.command;
 
 import chess.domain.manager.ChessManager;
 import chess.domain.manager.GameStatus;
