+++ src/test/java/chess/dao/ChessDaoTest.java
@@ -1,6 +1,6 @@
 package chess.dao;
 
-import chess.controller.dto.NewGameRequestDto;
+import chess.controller.dto.game.NewGameRequestDto;
 import chess.domain.manager.ChessManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
